requirements.txt
matplotlib~=3.7.1
numpy~=1.24.3
rasterio~=1.3.7
Pillow~=9.5.0
plotly~=5.14.1
pandas~=2.0.2
geopandas~=0.13.2
scipy~=1.10.1